add import and export links

+++ page-raceday-admin.php
@@ -8,12 +8,18 @@ if ( !current_user_can( 'manage_options' ) )  {
 
 global $BHAA;
 
-if(isset($_GET['action']))
-{
-	$runner = trim($_GET['runner']);
+if(isset($_GET['action'])){
 	$race = trim($_GET['raceid']);
-	error_log("delete action ".$runner.' '.$race);
-	$BHAA->registration->deleteRunner($runner,$race);
+	$booking = trim($_GET['booking']);
+	if($_GET['action']=='deleterunner') {
+		$runner = trim($_GET['runner']);
+		error_log("deleterunner ".$runner.' '.$race);
+		$BHAA->registration->deleteRunner($runner,$race);
+	} elseif($_GET['action']=='preregimport') {
+		error_log("preregimport ".$booking.' '.$race);
+	} elseif($_GET['action']=='preregexport') {
+		error_log("preregexport ".$booking.' '.$race);
+	}
 }
 
 get_header();
@@ -23,6 +29,10 @@ include_once 'page-raceday-header.php';
 $racetec = $BHAA->registration->listRegisteredRunners();
 
 echo '<h2>BHAA RACE DAY ADMIN</h2>';
+echo '<h3>Actions</h3>';
+echo sprintf('<h3><a href="/raceday-admin/?action=preregimport&booking=%d&raceid=%d">Import Pre Registered</a></h3>',113,2598);
+echo sprintf('<h3><a href="/raceday-admin/?action=preregexport&booking=%d&raceid=%d">Export Pre Registered</a></h3>',113,2598);
+echo '<hr/>';
 
 echo '<table id="raceteclist" >
 <tr class="row">
@@ -35,7 +45,7 @@ foreach($racetec as $racetec) : ?>
 <tr class="row">
 <td class="cell"><?php echo $racetec->firstname;?> <?php echo $racetec->lastname;?></td>
 <td class="cell"><?php echo $racetec->racenumber;?></td>
-<td class="cell"><?php echo sprintf('<a href="/raceday-admin/?action=delete&runner=%d&raceid=%d">%d</a>',$racetec->runner,$racetec->race,$racetec->runner);?></td>
+<td class="cell"><?php echo sprintf('<a href="/raceday-admin/?action=deleterunner&runner=%d&raceid=%d">%d</a>',$racetec->runner,$racetec->race,$racetec->runner);?></td>
 </tr>
 <?php endforeach;?>
 </table>
